[ANI-121-main-page] feat: msw 데이터 연동
