+++ apps/web/app/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return <div>메인 데이터 로딩 중...</div>
+}
